services/api/requirements.txt
setuptools>=69.0.2
wheel>=0.42.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
redis==5.0.1
motor==3.3.2
pandas>=2.1.4
numpy>=1.26.3
scikit-learn==1.4.0
transformers==4.36.2
torch==2.1.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.35
websockets==12.0
prometheus-client==0.19.0
pytest==7.4.4
httpx==0.24.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.0.0.post2
aiohttp==3.8.5
